Updated sitemap generation and added robots file
